requirements.txt
Flask>=2.0.3
flask-cors>= 4.0.0
google-cloud-datastore>=2.16.1
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
pyjwt>=2.8.0
werkzeug>=2.3.0
six>=1.16.0
jwt>=1.3.1
python-jose>=3.3.0